feat: header do not overrender , fixed Map elem
