+++ src/components/cities-list/cities-list.module.css
@@ -1,0 +1,5 @@
+.button__city {
+  border: none;
+  outline: none;
+  background-color: transparent;
+}
