+++ src/components/header/header.tsx
@@ -1,4 +1,4 @@
-import { Link } from 'react-router-dom';
+import { Link} from 'react-router-dom';
 import styles from './header.module.css';
 import { AppRoute, AuthStatus } from '../../const';
 import { useAppDispatch, useAppSelector } from '../../hooks';
@@ -17,12 +17,13 @@ function Header(): JSX.Element {
     }
   };
 
+
   return (
     <header className="header">
       <div className="container">
         <div className="header__wrapper">
           <div className="header__left">
-            <Link className="header__logo-link" to="/">
+            <Link className="header__logo-link" to={AppRoute.Root}>
               <img
                 className={`header__logo ${styles.header__logo}`}
                 src="img/logo.svg"
@@ -35,7 +36,7 @@ function Header(): JSX.Element {
             {isLoggedIn ?
               <ul className="header__nav-list">
                 <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to="#">
+                  <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favorites}>
                     <div className="header__avatar-wrapper user__avatar-wrapper">
                     </div>
                     <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
@@ -43,7 +44,7 @@ function Header(): JSX.Element {
                   </Link>
                 </li>
                 <li className="header__nav-item">
-                  <Link onClick={handleLogout} className="header__nav-link" to="#">
+                  <Link onClick={handleLogout} className="header__nav-link" to="/">
                     <span className="header__signout">Sign out</span>
                   </Link>
                 </li>
