+++ src/components/offers-on-map/offers-on-map.tsx
@@ -1,16 +1,10 @@
 import { OffersList } from '../../types/offers-list';
-import { useState , useEffect } from 'react';
+import { useState } from 'react';
 import Map from '../map/map';
 import Cards from '../cards/cards';
 import SortingOptions from '../sorting-options/sorting-options';
-import { useAppSelector , useAppDispatch } from '../../hooks';
+import { useAppSelector } from '../../hooks';
 import { getSortedOffers , getCityName } from '../../store/offers-data/offers-data.selectors';
-import { sortOffersByCity } from '../../store/offers-data/offers-data.slice';
-
-
-// type OffersOnMapProps = {
-//   selectedCityName: string;
-// }
 
 function OffersOnMap () : JSX.Element {
 
@@ -24,29 +18,26 @@ function OffersOnMap () : JSX.Element {
   };
 
   const selectedCityName = useAppSelector(getCityName);
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    dispatch(sortOffersByCity(selectedCityName));
-  }, [dispatch, selectedCityName]);
 
   return (
-    <div className="cities__places-container container">
-      <section className="cities__places places">
-        <h2 className="visually-hidden">Places</h2>
-        <b className="places__found">
-          {filteredOffersByCity.length} place{filteredOffersByCity.length !== 1 ? 's' : ''} to stay in {selectedCityName}
-        </b>
+    <div className="cities">
+      <div className="cities__places-container container">
+        <section className="cities__places places">
+          <h2 className="visually-hidden">Places</h2>
+          <b className="places__found">
+            {filteredOffersByCity.length} place{filteredOffersByCity.length !== 1 ? 's' : ''} to stay in {selectedCityName}
+          </b>
 
-        <SortingOptions />
+          <SortingOptions />
 
-        <Cards cardsList={filteredOffersByCity} onListItemHover={handleListItemHover}/>
+          <Cards cardsList={filteredOffersByCity} onListItemHover={handleListItemHover}/>
 
-      </section>
-      <div className="cities__right-section">
-        <section className="cities__map map">
-          {filteredOffersByCity.length !== 0 && <Map offers={filteredOffersByCity} selectedPoint={selectedOffer}/>}
         </section>
+        <div className="cities__right-section">
+          <section className="cities__map map">
+            {filteredOffersByCity.length !== 0 && <Map offers={filteredOffersByCity} selectedPoint={selectedOffer}/>}
+          </section>
+        </div>
       </div>
     </div>
   );
