+++ src/pages/main-screen/main-screen.tsx
@@ -1,4 +1,4 @@
-import CitiesList from '../../components/cities/cities';
+import CitiesList from '../../components/cities-list/cities-list';
 import Header from '../../components/header/header';
 import OffersOnMap from '../../components/offers-on-map/offers-on-map';
 
@@ -14,15 +14,13 @@ function MainScreen(): JSX.Element {
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
-          <section className="locations container">
 
-            <CitiesList />
+          <CitiesList />
 
-          </section>
         </div>
-        <div className="cities">
-          <OffersOnMap />
-        </div>
+
+        <OffersOnMap />
+
       </main>
     </div>
   );
